factor ClusterTotals up into DataFormats, updates to procompile
