+++ CommonTools/CandUtils/BuildFile.xml
@@ -1,6 +1,9 @@
 <use   name="DataFormats/Candidate"/>
 <use   name="DataFormats/TrackReco"/>
 <use   name="DataFormats/MuonReco"/>
+<use   name="DataFormats/VertexReco"/>
+<use   name="DataFormats/JetReco"/>
+<use   name="DataFormats/TauReco"/>
 <use   name="DataFormats/TrackerRecHit2D"/>
 <use   name="clhep"/>
 <export>
