+++ DataFormats/TrackerCommon/src/ClusterTotals.cc
@@ -1,0 +1,10 @@
+#include "DataFormats/TrackerCommon/interface/ClusterTotals.h"
+
+namespace reco { namespace utils {
+    ClusterTotals::ClusterTotals() : 
+      strip(0), 
+      pixel(0), 
+      stripdets(0), 
+      pixeldets(0) {
+    }       
+} }
